LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworkletglobalscope-sample-rate.https-expected.txt

CONSOLE MESSAGE: line 37: TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')

FAIL 
      Test sampleRate in AudioWorkletGlobalScope
     TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')
